# excourt_re/excourt_re-master/system/app/static/swagger.yaml
openapi: 3.0.0
info:
  title: ExCourt API
  description: API for court management system
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
...
  /exchangecourt/upload:
    post:
      summary: Upload exchange court request
      tags:
        - Exchange courts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - Exchange_uploader_id
                - Exchange_uploaded_court_id
              properties:
                Exchange_uploader_id:
                  type: integer
                  description: ID of the user uploading the request
                Exchange_uploaded_court_id:
                  type: integer
                  description: ID of the court for exchange
      responses:
        '201':
          description: Exchange request successfully uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Missing required fields
        '402':
          description: User has already uploaded a request for this court
        '500':
          description: Server error
  /exchangecourt/respond:
    post:
      summary: Respond to exchange court request
      tags:
        - Exchange courts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - Exchange_responser_id
                - Exchange_uploader_court_id
                - Exchange_responser_court_id
              properties:
                Exchange_responser_id:
                  type: integer
                  description: ID of the user responding to the request
                Exchange_uploader_court_id:
                  type: integer
                  description: ID of the court uploading the request
                Exchange_responser_court_id:
                  type: integer
                  description: ID of the court for exchange from the responder
      responses:
        '201':
          description: Exchange response successfully recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Missing required fields
        '402':
          description: Duplicate request detected
        '404':
          description: No exchange request found for the specified court
        '500':
          description: Server error
  /exchangecourt/get_response_records_by_student:
    post:
      summary: Get exchange response records for a student
      tags:
        - Exchange courts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - student_id
              properties:
                student_id:
                  type: integer
                  description: Student ID
      responses:
        '200':
          description: Exchange response records successfully retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        exchange_record_id:
                          type: integer
                        exchange_state:
                          type: string
                        exchange_uploader_court_id:
                          type: integer
                        exchange_responser_court_id:
                          type: integer
                        exchange_uploader_score:
                          type: integer
                          nullable: true
                        exchange_responser_score:
                          type: integer
                          nullable: true
                        exchange_responser_id:
                          type: integer
                        responser_info:
                          type: object
                          properties:
                            name:
                              type: string
                            profileurl:
                              type: string
                            nickname:
                              type: string
                            credit:
                              type: integer
                            level:
                              type: integer
                  status:
                    type: string
        '400':
          description: Missing student_id
        '500':
          description: Server error
  /exchangecourt/complete_exchange:
    post:
      summary: Complete exchange courts
      tags:
        - Exchange courts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - Exchange_uploader_court_id
                - Exchange_responser_court_id
                - Exchange_uploader_id
                - Exchange_responser_id
              properties:
                Exchange_uploader_court_id:
                  type: integer
                  description: ID of the court uploading the request
                Exchange_responser_court_id:
                  type: integer
                  description: ID of the court for exchange from the responder
                Exchange_uploader_id:
                  type: integer
                  description: ID of the user uploading the request
                Exchange_responser_id:
                  type: integer
                  description: ID of the user responding to the request
      responses:
        '200':
          description: Exchange successfully completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
        '400':
          description: Missing required fields
        '404':
          description: No matching exchange records found
        '500':
          description: Server error
  /exchangecourt/refuse:
    post:
      summary: Refuse exchange court request
      tags:
        - Exchange courts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - Exchange_uploader_court_id
                - Exchange_responser_court_id
                - Exchange_uploader_id
                - Exchange_responser_id
              properties:
                Exchange_uploader_court_id:
                  type: integer
                  description: ID of the court uploading the request
                Exchange_responser_court_id:
                  type: integer
                  description: ID of the court for exchange from the responder
                Exchange_uploader_id:
                  type: integer
                  description: ID of the user uploading the request
                Exchange_responser_id:
                  type: integer
                  description: ID of the user responding to the request
      responses:
        '200':
          description: Exchange request successfully refused
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
        '400':
          description: Missing required fields
        '404':
          description: No matching exchange records found
        '500':
          description: Server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: [] 